the suggestion of previous code review applied -> security checks in header, list_heritage and list_post are removed, check by if statements using a boolean variable (isAuthorized) is used.
recommend button issue ->the extra navbar element in header.jsp is removed, the button is in navbar in which login and profile buttons appear. it will not glitch again.

logo navigation -> when an unregistered user hits the logo, she is navigated to show_heritages instead of feed, which does not exist for such a user. when a registered user hits the logo, she is navigated to feed as usual.

profile page backend implementation -> implementation in profile_page branch is corrected, it is generic for all users. (navigate by username such as profile/salman)